Move circle_detector.h to include
