+++ rct_image_tools/include/rct_image_tools/circle_detector.h
