+++ rct_image_tools/src/rct_image_tools/circle_detector.cpp
@@ -46,7 +46,7 @@ the use of this software, even if advised of the possibility of such damage.
 // Slight Modification of OpenCV function to use ellipse fitting rather than
 // center of mass of contour to provide the location of the circle.
 
-#include "circle_detector.h"
+#include "rct_image_tools/circle_detector.h"
 
 #include <algorithm>
 #include <iterator>
